backend/requirements.txt
Flask==2.0.2
paho-mqtt==1.6.1
RPi.GPIO==0.7.0
